Added horizon leveling, and horizon locking
You can now press H to level the horizon, or shift-H to lock the horizon to a level position. Note that level the horizon is jittery when the camera is aligned with the x-axis. Also, I need to clean up the horizon-leveling code at some point; it was surprisingly hard to get it working properly.