Added Gorilla (GUI); added crosshair; broke -Wall -Werror
I've hooked in the Gorrila GUI system, and used it to add a crosshair to the app. Unfortunately, Gorilla doesn't compile with -Wall -Werror, so I had to turn that off for now.

Also, I did have to change a "None" inside an enum to "NONE", in order to get it to compile.